feat(refactor_integrity_checks): cleanup
