+++ src/Utils/ServiceCollectionExtensions/ServiceCollectionExtensions.cs
@@ -50,11 +50,10 @@ using form_builder.Services.ValidateService;
 using form_builder.TagParsers;
 using form_builder.TagParsers.Formatters;
 using form_builder.Validators;
-using form_builder.Validators.FormSchema.IntegrityChecks.FormSchema;
 using form_builder.Validators.IntegrityChecks;
 using form_builder.Validators.IntegrityChecks.Behaviours;
 using form_builder.Validators.IntegrityChecks.Elements;
-using form_builder.Validators.IntegrityChecks.FormSchema;
+using form_builder.Validators.IntegrityChecks.Form;
 using form_builder.Workflows.ActionsWorkflow;
 using form_builder.Workflows.DocumentWorkflow;
 using form_builder.Workflows.PaymentWorkflow;
