--- src/Validators/FormSchema/Class.cs
@@ -1,11 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace form_builder.Validators.FormSchema
-{
-    public class Class
-    {
-    }
-}
