--- src/Validators/FormSchema/IFormSchemaIntegrityValidator.cs
@@ -1,10 +1,0 @@
-using System.Threading.Tasks;
-using form_builder.Models;
-
-namespace form_builder.Validators
-{
-    public interface IFormSchemaIntegrityValidator
-    {
-        Task Validate(FormSchema schema);
-    }
-}
