--- src/Validators/FormSchema/IntegrityChecks/Elements/IElementSchemaIntegrityCheck.cs
@@ -1,11 +1,0 @@
-using form_builder.Models.Elements;
-using System.Threading.Tasks;
-
-namespace form_builder.Validators.IntegrityChecks.Elements
-{
-    public interface IElementSchemaIntegrityCheck
-    {
-        IntegrityCheckResult Validate(IElement element);
-        Task<IntegrityCheckResult> ValidateAsync(IElement element);
-    }
-}
