+++ src/Validators/IntegrityChecks/Behaviours/IBehaviourSchemaIntegrityCheck.cs
