+++ src/Validators/IntegrityChecks/Elements/IElementSchemaIntegrityCheck.cs
@@ -1,0 +1,11 @@
+using System.Threading.Tasks;
+using form_builder.Models.Elements;
+
+namespace form_builder.Validators.IntegrityChecks.Elements
+{
+    public interface IElementSchemaIntegrityCheck
+    {
+        IntegrityCheckResult Validate(IElement element);
+        Task<IntegrityCheckResult> ValidateAsync(IElement element);
+    }
+}
