+++ src/Validators/IntegrityChecks/Form/BookingFormCheck.cs
@@ -1,0 +1,37 @@
+﻿using form_builder.Constants;
+using form_builder.Enum;
+using System;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace form_builder.Validators.IntegrityChecks.Form
+{
+    public class BookingFormCheck : IFormSchemaIntegrityCheck
+    {
+        public IntegrityCheckResult Validate(FormSchema schema)
+        {
+            var integrityCheckResult = new IntegrityCheckResult();
+
+            if (!schema.Pages.Any(page => page.Elements.Any(element => element.Type.Equals(EElementType.Booking))))
+                return integrityCheckResult;
+
+            if (!schema.Pages.Any(page => page.PageSlug.Equals(BookingConstants.NO_APPOINTMENT_AVAILABLE, StringComparison.OrdinalIgnoreCase)))
+                integrityCheckResult.AddFailureMessage($"Booking Element Check, Form contains booking element, but is missing required page with slug {BookingConstants.NO_APPOINTMENT_AVAILABLE}.");
+
+            var bookingElements = schema.Pages.SelectMany(page => page.Elements.Where(element => element.Type.Equals(EElementType.Booking)));
+            var requiredFields = new[] { "customer.firstname", "customer.lastname" };
+            foreach (var requiredField in requiredFields)
+            {
+                if (!bookingElements.Any(
+                    element => element.Properties is not null && 
+                    element.Properties.TargetMapping is not null && 
+                    element.Properties.TargetMapping.Equals(requiredField, StringComparison.OrdinalIgnoreCase)))
+                integrityCheckResult.AddFailureMessage($"Booking Element Check, Booking element requires {requiredField} elements for reservation.");
+            }
+
+            return integrityCheckResult;
+        }
+
+        public async Task<IntegrityCheckResult> ValidateAsync(FormSchema schema) => await Task.Run(() => Validate(schema));
+    }
+}
