+++ src/Validators/IntegrityChecks/Form/DocumentDownloadCheck.cs
@@ -1,0 +1,30 @@
+using System.Linq;
+using System.Threading.Tasks;
+using form_builder.Enum;
+
+namespace form_builder.Validators.IntegrityChecks.Form
+{
+    public class DocumentDownloadCheck: IFormSchemaIntegrityCheck
+    {
+        public IntegrityCheckResult Validate(Models.FormSchema schema)
+        {
+            var integrityCheckResult = new IntegrityCheckResult();
+
+            if (!schema.DocumentDownload)
+                return IntegrityCheckResult.ValidResult;
+
+            if (schema.DocumentType.Any() && schema.DocumentType.Any(_ => _ == EDocumentType.Unknown))
+            {
+                integrityCheckResult.AddFailureMessage($"Document Download Check, Unknown document download type configured for form '{ schema.FormName }'");
+            }
+            else
+            {
+                integrityCheckResult.AddFailureMessage($"Document Download Check, No document download type configured.");
+            }
+
+            return integrityCheckResult;
+        }
+
+        public async Task<IntegrityCheckResult> ValidateAsync(Models.FormSchema schema) => await Task.Run(() => Validate(schema));
+    }
+}
