+++ src/Validators/IntegrityChecks/Form/IFormSchemaIntegrityCheck.cs
@@ -1,0 +1,10 @@
+using System.Threading.Tasks;
+
+namespace form_builder.Validators.IntegrityChecks.Form
+{
+    public interface IFormSchemaIntegrityCheck
+    {
+        IntegrityCheckResult Validate(Models.FormSchema schema);
+        Task<IntegrityCheckResult> ValidateAsync(Models.FormSchema schema);
+    }
+}
