+++ src/Validators/IntegrityChecks/IFormSchemaIntegrityValidator.cs
@@ -1,0 +1,10 @@
+using System.Threading.Tasks;
+using form_builder.Models;
+
+namespace form_builder.Validators.IntegrityChecks
+{
+    public interface IFormSchemaIntegrityValidator
+    {
+        Task Validate(FormSchema schema);
+    }
+}
