+++ src/Validators/IntegrityChecks/IntegrityCheckResult.cs
